refactor: move the registry surface-key derivation into src/
workers/registry-sync-api.ts calls itself the ONLY write path into the
registry database and nothing calls it -- its callers were GitHub
Actions scripts no workflow invokes, and surface_history has been
frozen since 2026-08-02 (#9779). Re-homing that sync onto a Worker
means the Worker needs surface_key, and a Worker cannot import
scripts/lib.ts: that module pulls in node:fs, node:child_process and
undici at the top level.

Re-implementing the derivation is not an option. surface_key is the
locator joining a surface to its probe history, so a normaliser that
differs on one url mints a key matching nothing -- the surface looks
new, its history looks orphaned, and neither raises an error.

The blocker turned out not to be one. The chain bottoms out in
node:net's isIP and BlockList, which I had assumed were unavailable in
Workers; node:net is fully supported under nodejs_compat and
Cloudflare's 2025-01-28 changelog names net.BlockList explicitly. So
the functions move as they are.

MOVED, NOT REWRITTEN: scripts/lib.ts re-exports every name, so there is
exactly one implementation and no caller changes. The first extraction
attempt sliced on guessed line ranges and pulled in the tails of two
neighbouring functions; redone with top-level brace boundaries.

Proven against production rather than a fixture: deriving the key for
all 3,444 live surfaces and comparing to the stored value gives 3,444
matches and zero mismatches. 411 tests across the seven suites covering
these functions still pass.

Closes #10028
